For consistency put test with others
Whether this is right or wrong is a matter of opinion, but having the tests
together makes more sense probably
